northd: fix stateless nat.
ip.dst is unnecessary in the action and it could generate problem in the
following scenario:
- Connection between the router from client and server has multiple paths (ECMP)
and multiple chassis (ovn-gw1 and ovn-gw2).
- LB is distributed between differents AZs with the same backends configured.
Client send a FIN/ACK by ovn-gw1, server replies the FIN/ACK and waits
for the last ACK.
If the client send the last ACK by ovn-gw2, backend could not receive
because ovn-gw2 applies ip4.dst in the first packet, and if it the first
backend configured was not chosen before, this last ACK never is delivered to
the correct backend.

Signed-off-by: Lucas Vargas Dias <lucas.vdias@luizalabs.com>